Did a bunch of stuff:
Made the website be (almost) uniformly dark mode

Made it so you can View Albums

Added the framework for viewing songs

Made it so you can preview the playlist when searching songs, and when you add a song to the playlist the preview refreshes and shows your changes.

Made it so you can remove songs from playlists

Made the images keep a square shape when they resize.

Made it so if the connection is closed it refreshes the page (except for the main page, will do soon maybe)